+++ pgcopylib/__init__.py
@@ -33,4 +33,4 @@ __all__ = (
     "PGOidToDType",
 )
 __author__ = "0xMihalich"
-__version__ = "0.1.2"
+__version__ = "0.1.3"